add multiple dots
