+++ src/components/styles/Layout.scss
@@ -20,6 +20,7 @@
   z-index: 2;
   width: 100%;
   height: auto;
+  opacity: 0;
 }
 
 .text-wrapper {
