+++ src/utils/constants.ts
@@ -1,5 +1,4 @@
 // SCENE 2 CONSTANTS
-
 export const CLOUD_WIDTH_MIN = 100;
 export const CLOUD_WIDTH_MAX = 120;
 export const CLOUD_HEIGHT_MIN = 60;
@@ -8,10 +7,17 @@ export const CLOUD_HEIGHT_MAX = 80;
 export const NUM_CLOUD_SECTION_PARTICLES = 150;
 export const PARTICLE_DIAMETER = 15;
 
+export const CLOUD_OFFSETS = [
+  {x: -1 * window.innerWidth / 3, y: -1 * window.innerWidth / 14 },
+  {x: -1 * window.innerWidth / 8, y:      window.innerWidth / 14 },
+  {x:      window.innerWidth / 8, y: -1 * window.innerWidth /  9 },
+  {x:      window.innerWidth / 3, y:                           0 },
+];
+
 export const CLOUD_SECTION_OFFSETS = [
-  {x: -80, y: 0},     // left
-  {x: 80, y: 0},     // right
-  {x: 30, y: 30},     // below
+  {x: -80, y:   0},     // left
+  {x:  80, y:   0},     // right
+  {x:  30, y:  30},     // below
   {x: -20, y: -50},     // above
 ];
